16-ML/ML_Ops/ml_project/requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
boto3==1.24.59
botocore==1.27.59
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
cryptography==38.0.3
cycler==0.11.0
databricks-cli==0.17.3
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.0
dpath==2.0.6
dulwich==0.20.50
dvc==2.33.1
dvc-data==0.25.2
dvc-http==2.27.2
dvc-objects==0.12.2
dvc-render==0.0.12
dvc-s3==2.20.1
dvc-task==0.1.4
dvclive==0.12.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
grandalf==0.6
greenlet==2.0.0
gunicorn==20.1.0
htmlmin==0.1.12
hydra-core==1.2.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iterative-telemetry==0.0.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib==3.5.3
matplotlib-inline==0.1.6
missingno==0.5.1
mistune==2.0.4
mlflow==1.30.0
multidict==6.0.2
multimethod==1.9
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
pandas-profiling==3.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.31
protobuf==4.21.9
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
pygit2==1.10.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
querystring-parser==1.2.4
requests==2.28.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2022.10.0
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
scmrepo==0.1.2
seaborn==0.12.1
Send2Trash==1.8.0
shortuuid==1.0.9
shtab==1.5.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
stack-data==0.6.0
statsmodels==0.13.4
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
visions==0.7.5
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
zc.lockfile==2.0
zipp==3.10.0
